Extended menu column component to accept static list of actions
